requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
blinker==1.9.0
blockbuster==1.5.25
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
cryptography==44.0.3
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
executing==2.2.1
forbiddenfruit==0.1.4
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-tools==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonschema_rs==0.29.1
langchain==1.0.3
langchain-anthropic==1.0.1
langchain-core==1.0.3
langchain-openai==1.0.2
langgraph==1.0.2
langgraph-api==0.5.5
langgraph-checkpoint==3.0.1
langgraph-cli==0.4.7
langgraph-prebuilt==1.0.2
langgraph-runtime-inmem==0.16.0
langgraph-sdk==0.2.9
langsmith==0.4.41
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
narwhals==2.10.2
numpy==2.3.4
openai==2.7.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
prompt_toolkit==3.0.52
protobuf==6.33.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.28.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.50.0
streamlit==1.51.0
structlog==25.5.0
tenacity==9.1.2
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.14
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
